requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-groq>=0.0.3
python-dotenv>=0.19.0 